+++ server/include/CmdParser.h
@@ -1,0 +1,3 @@
+#pragma once
+#include <boost/program_options.hpp>
+boost::program_options::variables_map parseCommandLine(int argc, char* argv[]);