+++ server/src/main.cpp
@@ -1,22 +1,30 @@
 #include <iostream>
+#include <boost/asio.hpp>
 #include <TcpServer.h>
 #include "Database.h"
+#include "CmdParser.h"
 
 void registerServerRoutes();
 void startAutoRefreshTimer(boost::asio::io_context& io_context);
+
 int main(int argc, char* argv[]) {
-    std::cout << "Server Started..." << std::endl;
+    auto vm = parseCommandLine(argc, argv);
+    uint16_t port = vm["port"].as<uint16_t>();
+    std::cout << "Server Starting on port " << port << "..." << std::endl;
+
     try {
         Database::getInstance();
         std::cout << "Database initialized" << std::endl;
         registerServerRoutes();
 
         boost::asio::io_context io_context;
-        TcpServer server(io_context, 8080);
+        TcpServer server(io_context, port);
+
         startAutoRefreshTimer(io_context);
         io_context.run();
     } catch (std::exception& e) {
-        std::cerr << e.what() << std::endl;
+        std::cerr << "System Error: " << e.what() << std::endl;
+        return 1;
     }
     return 0;
 }