+++ changelog.md
@@ -1,3 +1,3 @@
-* Fix labels on linux using system colors not app specific colors
-* Add Unit testing and tests module
-* Release for Milestone 2
+* Add public server and add it natively to the client
+* Add --port argument for server
+* Update default custom server port to match server default port